Security fixes.

+++ src/types/dc6.c
@@ -40,7 +40,7 @@ dc6_t *dc6_load(const char *path) {
 void dc6_decode_frame(const dc6_t *dc6, uint32_t frame_index) {
     dc6_frame_t *frame = &dc6->frames[frame_index];
 
-    frame->indexed_pixel_data = malloc(frame->header.width * frame->header.height);
+    frame->indexed_pixel_data = malloc((size_t)frame->header.width * frame->header.height);
     FAIL_IF_NULL(frame->indexed_pixel_data);
     memset(frame->indexed_pixel_data, 0, (size_t)frame->header.width * frame->header.height);
